+++ Release-Notes.md
@@ -1,5 +1,11 @@
 # FarNet.FSharp.PowerShell Release Notes
 
+## v0.0.3
+
+Make getting types lazy, they are only used on `Get-Type`.
+
+`*Invoke2()`: if types do not match, use `LanguagePrimitives.ConvertTo()`.
+
 ## v0.0.2
 
 First preview.